+++ src/main/java/duke/DukeException.java
@@ -1,7 +1,8 @@
+package duke;
+
 public class DukeException extends Exception{
 
     protected static final String EMPTY_INPUT = "Please enter something!";
-    protected static final String EMPTY_NAME = "Please give a name for the task";
     protected static final String UNKNOWN_COMMAND = "I don't know what that means";
     protected static final String INVALID_INDEX = "The task number is invalid";
     protected static final String INVALID_FORMAT = "The task format is invalid";
