+++ src/main/java/ui/Ui.java
@@ -1,3 +1,6 @@
+package ui;
+
+import task.TaskList;
 import java.util.Scanner;
 
 public class Ui {
